OOPlize the functions, makes the main function look much "neater"
