+++ duke.txt
@@ -1,3 +1,1 @@
-D | 0 | todo | 2020-11-02
-T | 0 | hello
-T | 0 | hjhjhj
+E | 0 | jskajdf | 2020-09-12
