+++ src/main/java/Duke/Exceptions/ipException.java
@@ -1,7 +1,7 @@
-package Duke.Exceptions;
+package duke.exceptions;
 
-import Duke.Duke;
+import duke.Duke;
 
-public class ipException extends Exception {
+public class IpException extends Exception {
 
 }
