+++ src/main/java/Duke/Task/Deadline.java
@@ -1,7 +1,5 @@
-package Duke.Task;
+package duke.task;
 
-import Duke.Duke;
-import Duke.Task.Task;
 
 public class Deadline extends Task {
     protected String time;
