+++ src/main/java/Duke/Task/Event.java
@@ -1,4 +1,4 @@
-package Duke.Task;
+package duke.task;
 
 public class Event extends Task {
     protected String time;
