+++ src/main/java/Duke/Task/ToDo.java
@@ -1,4 +1,4 @@
-package Duke.Task;
+package duke.task;
 
 public class ToDo extends Task {
     protected String time;
